Enable use of local build of Slang
For development, it is annoying to have to use a version of Slang pulled from packman.

This change makes it so that we can use a local copy of Slang checked out under `Framework\Externals\slang`, and it will automatically use the Debug x64 build instead of the release one, so we can actually debug issues.

This change should *not* get merged into upstream Falcor.