Enable use of Slang IR
This change makes Falcor use the Slang IR-based code generator, which is needed to take advantage of any of the advanced features of the language (generics, `interface`s, etc.).

User code written in HLSL/GLSL will continue to use the existing source-to-source "rewriter" path, so this will ideally only affect writers of library code.