Fix: don't use #import
The `NormalMapFiltering` sample was using `#import`, but this seems to confuse things because some files end up getting imported multiple times, which breaks reflection information.

There is no reason for the sample to use `#import` any more, and in fact we should deprecate and remove the feature as soon as possible.