Fix: use Slang compile-time loops for unrolling
The IR-based Slang code generator produces HLSL loops in a form that the HLSL compiler can't unroll, so that it complains if they have the `[unroll]` attribute and/or are required to be unrolled because they index things that aren't dynamically indexable.

The quick-fix solution here is to use the Slang `$for` construct for compile-time looping.

A longer term solution would be to have Slang implement unrolling in its IR. This is a major chunk of work, though, so we are putting it off as long as possible.