Change `FeatureDemoCommon.{hlsli -> slang}`
This sample was actually doing something semantically shady in Slang code: two different `.slang` files in the same compile request were both `#include`ing the same `cbuffer` declaration.

When that is done by two `.glsl` or two `.hlsl` files it is supposed to be allowed, because the two declarations are treated as distinct declaratiosn of the same parameter, in different translation units (like two different `extern` declaratiosn of the same global in C/C++).

All the `.slang` files in a compile request belong to a single translation unit right now (maybe that should change...), so that these end up being two distinct global `cbuffer` that just happen to have the same name. The front end should diagnose this as an error, but it doesn't right now.

This results in mix-ups in IR type legalization that break the `FeatureDemo` shaders for the IR.