Refactor the way program versions get created
Right now user code assumes that it can create a `Program`, (optionally) set some `#define`s, and then extract a `ProgramVersion` with `getActiveVersion` to get both:

- Reflection information, suitable to use with any similar/compatible versions
- Actual bind-able shader kernels, that can go into a PSO

The problem that arises here is that in some cases (e.g., when adding "shader components" support) we need to use information from the `ProgramVars` to inform the generation of bind-able shader kernels. For example, if we have a global-scope type parameter to stand in for a material type, then we can't generate final code until we plug in a concrete type for that parameter.

Trying to create a `ProgramVersion` without that additional info can currently lead to compilation failure ("you didn't tell me a concrete type here, so what code am I supposed to generate"). However, we currently can't create a `ProgamVars` (or the related reflection data) without creating a `ProgramVersion` first.

The root of the problem is that we don't want a single operation that both gets/creates reflection information *and* compiled kernel code, but rather we want thse to be two different steps.
This change tries to refactor the design along those lines, without changing the common case for user code (so that application code doesn't need to change).

Notes:

- Most of what `ProgramVersion` used to hold is now a new `ProgramKernels` type, that just holds the final `Shader` objects
  - Graphics and compute state objects now hold `ProgramKernels` instead of `ProgramVersion`s

- A new `ProgramVersion` type is created that is conceptually a "snapshot" of a `Program` given the `#define`s that were active when the snapshot was taken.

- `Program::getActiveVersion` applies the Slang compiler to the input code (given the active defines), but explicitly does *not* extract compiled kernel code. It passes on the resulting reflection data to create a `ProgramVersion`.
  - Note: the Slang API still receives exactly the same input right now, so the error cases are still errors with just this change.

- Currently there is a bottleneck (in `GraphicsState::getPSO` and `ComputeState::getGSO`) where the actual kernels are needed (because the state objects store the kernels), and it just so happens that we have a `ProgramVars` available in each of those cases (this is not coincidental). We can thus invoke a new operation at that point that takes into account the `ProgramVersion` (which remembers the `Program` and the `#define`s) *and* the `ProgramVars` and generates final kernel code.

- In order to minimize duplication, both of these compile steps share a lot of the Slang API calls via subroutines.

- There is currently no attempt made to actually use information from the `ProgramVars`, and the "caching" logic in `ProgramVersion` for re-using `ProgramKernels` is stubbed out (there is only one cached set of kernels).