Pick separate compilation profiles for Vulkan
Currently, Falcor passes HLSL-specific profile names (e.g., `vs_5_0`) when compiling for Vulkan, which means that Slang can't figure out an appropriate GLSL version to declare when cross-compiling. This change makes it so that we use Slang's default GLSL target profiles, which currently amount to `#version 450`.

This might be something Slang can improve by providing a mapping from HLSL profiles to GLSL equivalents, but that is hard to define precisely.